requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
configparser==3.8.1
dj-database-url==0.5.0
Django==3.1
django-crispy-forms==1.11.0
djangorestframework==3.11.1
future==0.18.2
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
jsonschema==2.6.0
Pillow==7.2.0
psycopg2==2.8.6
Pygments==2.7.1
PySocks==1.7.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.1
sqlparse==0.3.1
ua-parser==0.10.0
urllib3==1.25.10
user-agents==2.2.0
whitenoise==5.2.0